learn_python/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.1
Brotli==1.1.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
cloudinary==1.44.1
colorama==0.4.6
distro==1.9.0
Django==5.2.6
django-allauth==65.11.2
django-bootstrap5==25.2
django-cloudinary-storage==0.3.0
django-jazzmin==3.0.1
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.27.2
idna==3.10
packaging==25.0
psycopg2-binary==2.9.10
pydantic==2.11.9
pydantic_core==2.33.2
Pygments==2.19.2
python-decouple==3.8
python-dotenv==1.1.1
requests==2.32.5
six==1.17.0
sniffio==1.3.1
spitch==1.34.0
sqlparse==0.5.3
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
whitenoise==6.10.0
